_code/pyproject.toml
[project]
name = "engram-r"
version = "0.7.0"
description = "Co-scientist EngramR for Obsidian -- generate-debate-evolve hypothesis architecture"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "scipy>=1.14",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.5",
]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "ruff>=0.11",
    "black>=25.0",
]
bot = [
    "slack-bolt>=1.22",
    "anthropic>=0.70",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.black]
line-length = 88
target-version = ["py311"]
